fix: read NL recent from Parlamint v4
