+++ backend/corpora/parliament/utils/parlamint_v4.py
@@ -1,6 +1,20 @@
+from glob import glob
+from string import punctuation
+from typing import Iterable
+
 from addcorpus.python_corpora.extract import XML, Combined, Metadata
-from bs4.element import NavigableString
-import datetime
+from ianalyzer_readers.xml_tag import Tag
+from bs4.element import NavigableString, Tag as Node
+from bs4 import BeautifulSoup
+
+from addcorpus.es_mappings import non_indexed_text_mapping, keyword_mapping
+from addcorpus.python_corpora.corpus import FieldDefinition
+from addcorpus.python_corpora.filters import MultipleChoiceFilter
+from corpora.parliament.utils.parlamint import (
+    metadata_attribute_transform_func,
+    person_attribute_extractor,
+)
+
 
 """
 This file was created as an updated utils file for the ParlaMint dataset, version 4.0. The previous utils file
@@ -62,6 +76,23 @@ def extract_org_data(node):
         'political_orientation': political_orientation
     }
 
+
+def extract_soup(filename: str) -> Node:
+    with open(filename, 'rb') as f:
+        soup = BeautifulSoup(f.read(), "xml")
+    return soup
+
+
+def load_people_data(data_directory: str) -> Node:
+    person_file = glob(f'{data_directory}/*-listPerson.xml')[0]
+    return extract_soup(person_file)
+
+
+def load_party_data(data_directory: str) -> Node:
+    party_file = glob(f'{data_directory}/*-listOrg.xml')[0]
+    return extract_soup(party_file)
+
+
 def extract_all_org_data(soup):
     orgs_list = soup.find('listOrg')
     org_data = map(extract_org_data, orgs_list.find_all('org'))
@@ -112,37 +143,6 @@ def extract_people_data(soup):
     return {
        person['id']: person for person in person_data
     }
-
-def extract_role_data(soup):
-    role_nodes = soup.find('encodingDesc').find_all('category')
-    # return dict that maps IDs to terms data contains duplicate role IDs
-    # go through data in reverse order so earlier (more general) terms 
-    # overwrite later (more specific) ones
-    return {
-        node['xml:id']: node.find('term').text.strip()
-        for node in reversed(role_nodes)
-    }
-
-def metadata_attribute_transform_func(attribute):
-    """
-    Creates a transformation function that extracts and cleans a specific 
-    attribute from a collection.
-    """
-    def get_attribute(which, collection):
-        if which and collection and which in collection:
-            value = collection[which][attribute]
-            return clean_value(value)
-
-    return lambda values: get_attribute(*values)
-
-def person_attribute_extractor(attribute, id_attribute = 'who'):
-    """Extractor that finds the speaker ID and returns one of the person's
-    attributes defined in extract_person_data()"""
-    return Combined(
-        XML(attribute=id_attribute),
-        Metadata('persons'),
-        transform = metadata_attribute_transform_func(attribute),
-    )
 
 def current_party_id_extractor():
     """Extractor that finds the current party, given a date
@@ -196,3 +196,127 @@ def transform_current_party_id(data):
         return party_nodes[-1].get('ref', 'NA')
     else:
         return current_parties[-1] #return the last org in the list since that is usually the most recent one.
+
+
+def ner_keyword_field(entity: str):
+    return FieldDefinition(
+        name=f"{entity}:ner-kw",
+        display_name=f"Named Entity: {entity.capitalize()}",
+        searchable=True,
+        es_mapping=keyword_mapping(enable_full_text_search=True),
+        search_filter=MultipleChoiceFilter(
+            description=f"Select only speeches which contain this {entity} entity",
+            option_count=100,
+        ),
+        extractor=Combined(
+            XML(attribute="xml:id"),
+            Metadata("ner"),
+            transform=lambda x: get_entity_list(x, entity),
+        ),
+    )
+
+
+def detokenize_parlamint(tokens: Iterable[str]) -> str:
+    """Detokenize the content of `w` and `pc` tags in the ParlaMint XML
+    The `join="right"` attribute indicates that there should not be whitespace after the word
+    """
+    output = ""
+    for token in tokens:
+        if token.get("join") != "right":
+            output += f"{token.string} "
+        else:
+            output += token.string
+    # do not include the last character (always whitespace) in the output
+    return output[:-1]
+
+
+def format_annotated_segment(element: Node) -> str:
+    """For each <seg> tag, extract the annotations indicated by <name>"""
+    annotations = element.find_all("name")
+    formatted_annotations = [format_annotated_text(anno) for anno in annotations]
+    return "".join(formatted_annotations)
+
+
+def format_annotated_text(element: Node) -> str:
+    """For each <name> tag, format the annotation,
+    and embed it in the text extracted from adjoining <w> and <pc> tags
+    """
+    output = ""
+    tokens = [el.extract() for el in element.find_previous_siblings(["w", "pc"])]
+    output += detokenize_parlamint(reversed(tokens))
+    annotated = element.find_all("w")
+    formatted = " ".join([a.string for a in annotated])
+    if output:
+        # if there is preceding text, add whitespace prior to annotation
+        output += " "
+    output += f"[{formatted}]({element['type']})"
+    if not element.find_next_sibling("name"):
+        # after last annotation, add remaining text
+        remaining_text = detokenize_parlamint(element.find_next_siblings(["w", "pc"]))
+        if remaining_text[0] not in punctuation:
+            # remaining text does not start with punctuation: add whitespace
+            output += " "
+        output += remaining_text
+    return output
+
+
+def speech_ner():
+    return FieldDefinition(
+        name="speech:ner",
+        hidden=True,
+        es_mapping=non_indexed_text_mapping(),
+        display_type="text_content",
+        searchable=True,
+        extractor=XML(
+            Tag("seg"),
+            multiple=True,
+            extract_soup_func=format_annotated_segment,
+            transform=lambda x: "\n".join(x),
+        ),
+    )
+
+
+def extract_speech(segment: Node) -> str:
+    text_nodes = segment.find_all(["w", "pc"])
+    return detokenize_parlamint(text_nodes)
+
+
+def get_entity_shorthand(entity: str):
+    if entity == "location":
+        return "LOC"
+    elif entity == "miscellaneous":
+        return "MISC"
+    elif entity == "organization":
+        return "ORG"
+    else:
+        return "PER"
+
+
+def get_entity_list(extracted_data: tuple[str, dict], entity: str) -> list[str]:
+    '''collect all named entities for the processed speech and this category
+
+    Parameters:
+        extracted_data: tuple of the speech id and the metadata dictionary
+        entity: string of the entity class (location /misc / organization / person)
+
+    '''
+    speech_id, metadata = extracted_data
+    shorthand = get_entity_shorthand(entity)
+    return list(set(metadata.get(speech_id).get(shorthand)))
+
+
+def extract_named_entities(xml_file: str) -> dict:
+    '''Extract the named entities from the xml file, and save them, ordered by speech id,
+    in a dictionary, which will be used to populate the NER keyword fields'''
+    with open(xml_file) as f:
+        soup = BeautifulSoup(f, 'xml')
+    speeches = soup.find_all("u")
+    output = dict()
+    for speech in speeches:
+        annotations_dict = {"LOC": list(), "MISC": list(), "ORG": list(), "PER": list()}
+        annotations = speech.find_all("name")
+        for annotation in annotations:
+            annotated = " ".join([word.string for word in annotation.find_all("w")])
+            annotations_dict[annotation["type"]].append(annotated)
+        output[speech["xml:id"]] = annotations_dict
+    return output
